Cette pull request introduit la création et l'implémentation de la table
`Marque` dans l'application. Elle vise à structurer et stocker les
informations relatives aux marques, améliorant ainsi la gestion des
données de notre application.

### gitignore
- Modification du git ignore pour ne pas envoyer l'upload d'images de
test.

### src/Controller/MainController.php
- **Rectification** : pas de modification au final dans
`MainController`, affichage de toutes les marques par un lien de
navigation.

### src/Controller/MarqueController.php
- Modification du Endpoint dans la méthode `index` pour afficher toutes
les marques pour l'utilisateur ( même si certaines informations sont
cachées et limiter qu'aux utilisateurs ayant le *ROLE_ADMIN* )

### src/Form/MarqueType.php
- Précision et sécurisation pour le formulaire, ajouter des contraintes
pour certains champs.
- `Nom`: contraintes pour ne pas être vide, longueur
- `Description` : **textarea** pour avoir de la place et que le champ
soit assez grand.
- `Logo` : **Input** pour insérer une image, limitations pour la taille
du fichier et les fichiers acceptés, messages si erreurs.
inputs date masqués pour **createdAt** et **updatedAt**.

### src/Service/UploadFileService.php
- Création du service pour gérer l'upload/suppression de fichier dans
l'application injecté dans `MainController`

### templates/main/index.html.twig
- Liens de navigation, pour afficher la liste des marques de
l'application.

### templates/marque/index.html.twig
- Gestion des rôles affichage limité pour **"ROLE_USER"**. Les actions,
métadonnées, sont affichées que pour les utilisateurs qui ont au minimum
un **ROLE_ADMIN**
- Affichage conditionnel pour les images : si une image affiche l'image
sinon message "Pas d'images"

### templates/marque/show.html.twig
- Gestion des rôles affichage limité pour **"ROLE_USER"**. Les actions,
métadonnées, sont affichées que pour les utilisateurs qui ont au minimum
un ROLE_ADMIN
Affichage conditionnel pour les images : si une image affiche l'image
sinon message "Pas d'images"

+++ src/Controller/MarqueController.php
@@ -5,10 +5,13 @@ namespace App\Controller;
 use App\Entity\Marque;
 use App\Form\MarqueType;
 use App\Repository\MarqueRepository;
+use App\Service\UploadFileService;
 use DateTimeImmutable;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bridge\Doctrine\Attribute\MapEntity;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\DependencyInjection\Attribute\Autowire;
+use Symfony\Component\HttpFoundation\File\File;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
@@ -18,7 +21,7 @@ use Symfony\Component\String\Slugger\SluggerInterface;
 // #[Route('/marque')]
 final class MarqueController extends AbstractController
 {
-    #[Route('/admin/marque', name: 'app_marque_index', methods: ['GET'])]
+    #[Route('/marque', name: 'app_marque_index', methods: ['GET'])]
     public function index(MarqueRepository $marqueRepository): Response
     {
         return $this->render('marque/index.html.twig', [
@@ -27,13 +30,39 @@ final class MarqueController extends AbstractController
     }
 
     #[Route('/admin/marque/new', name: 'app_marque_new', methods: ['GET', 'POST'])]
-    public function new(Request $request, EntityManagerInterface $entityManager, SluggerInterface $slugger): Response
-    {
+    public function new(
+        Request $request,
+        EntityManagerInterface $entityManager,
+        SluggerInterface $slugger,
+        #[Autowire('%kernel.project_dir%/public/uploads/marques')] string $marquesDirectory,
+        UploadFileService $uploadFileService
+    ): Response {
         $marque = new Marque();
         $form = $this->createForm(MarqueType::class, $marque);
         $form->handleRequest($request);
 
+
         if ($form->isSubmitted() && $form->isValid()) {
+
+            if ($marque) {
+                $logoFile = $form->get('logo')->getData();
+
+                if ($logoFile) {
+
+                    // $originalLogoFilename = pathinfo($logoFile->getClientOriginalName(), PATHINFO_FILENAME);
+                    // $safeFileName = $slugger->slug($originalLogoFilename);
+                    // $newLogoFileName = $safeFileName . '-' . uniqid() . '.' . $logoFile->guessExtension();
+                    $fileToSend = $uploadFileService->uploadFile($logoFile);
+                    // dd($newLogoFileName);
+                    if ($fileToSend) {
+                        // $logoFile->move($marquesDirectory, $newLogoFileName);
+                        // $marque->setLogo($newLogoFileName);
+                        $logoFile->move($marquesDirectory, $fileToSend);
+                        $marque->setLogo($fileToSend);
+                    }
+                }
+            }
+
             $nomCategorie = $marque->getNom();
             $nomCategorieSlug = $slugger->slug($nomCategorie);
             $marque->setSlug(strtolower(($nomCategorieSlug)));
@@ -58,16 +87,51 @@ final class MarqueController extends AbstractController
     }
 
     #[Route('admin/marque/{slug}/edit', name: 'app_marque_edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request, #[MapEntity(mapping: ['slug' => 'slug'])] Marque $marque, EntityManagerInterface $entityManager, SluggerInterface $slugger): Response
-    {
+    public function edit(
+        Request $request,
+        #[MapEntity(mapping: ['slug' => 'slug'])] Marque $marque,
+        EntityManagerInterface $entityManager,
+        SluggerInterface $slugger,
+        #[Autowire('%kernel.project_dir%/public/uploads/marques')] string $marquesDirectory,
+        UploadFileService $uploadFileService
+    ): Response {
+        $logoMarque = $marque->getLogo();
         $form = $this->createForm(MarqueType::class, $marque);
         $form->handleRequest($request);
+        // $pathInPublic = 'uploads/marques/';
 
         if ($form->isSubmitted() && $form->isValid()) {
+
+            $logoFile = $form->get('logo')->getData();
+            // dd($logoFile);
+            if ($logoFile) {
+                $originalLogoFilename = pathinfo(
+                    $logoFile->getClientOriginalName(),
+                    PATHINFO_FILENAME
+                );
+
+                // $safeFileName = $slugger->slug($originalLogoFilename);
+                // $newLogoFileName = $safeFileName . '-' . uniqid() . '.' . $logoFile->guessExtension();
+                $fileToSend = $uploadFileService->uploadFile($logoFile);
+                // dd($fileToSend);
+                if ($fileToSend) {
+
+                    $logoFile->move($marquesDirectory, $fileToSend);
+                    $marque->setLogo($fileToSend);
+                    $marqueFilePath = 'uploads/marques/' . $logoMarque;
+                    // dd($marqueFilePath);
+                    if ($logoMarque) {
+                        unlink($marqueFilePath);
+                    }
+                }
+            } else {
+                $marque->setLogo($logoMarque);
+            }
             $nomMarque = $form->get('nom')->getData();
             $nomSlugMarque = $slugger->slug($nomMarque);
-            $marque->setSlug($nomSlugMarque);
+            $marque->setSlug(strtolower($nomSlugMarque));
             $marque->setUpdatedAt(new DateTimeImmutable());
+
             $entityManager->persist($marque);
             $entityManager->flush();
 
@@ -81,11 +145,17 @@ final class MarqueController extends AbstractController
     }
 
     #[Route('/admin/marque/{id}', name: 'app_marque_delete', methods: ['POST'])]
-    public function delete(Request $request, Marque $marque, EntityManagerInterface $entityManager): Response
+    public function delete(Request $request, Marque $marque, EntityManagerInterface $entityManager, UploadFileService $uploadFileService): Response
     {
         if ($this->isCsrfTokenValid('delete' . $marque->getId(), $request->getPayload()->getString('_token'))) {
-            $entityManager->remove($marque);
-            $entityManager->flush();
+            $marqueFilePath = 'uploads/marques/' . $marque->getLogo();
+            $fileToRemove = $uploadFileService->removeFile($marqueFilePath);
+            if ($fileToRemove) {
+
+                $entityManager->remove($marque);
+                $entityManager->flush();
+            }
+            $this->addFlash('fail', 'Le fichier ne s\'est pas supprimé correctement');
         }
 
         return $this->redirectToRoute('app_marque_index', [], Response::HTTP_SEE_OTHER);
